NumeralsConverter interface with ArabicToRomanNumeralsConverter implementation of converter arabic numbers from 1 to 3999 inclusive have been finallized; ArabicToRomanNumeralsConverterAllTests class that is a test suit with ArabicToRomanNumeralsConverterNegativeTests class and ArabicToRomanNumeralsConverterParametrisedTest class have been completed.
