+++ TDDFeb2018/src/com/epam/tdd/lesson180216/Model.java
@@ -1,0 +1,13 @@
+package com.epam.tdd.lesson180216;
+
+public class Model {
+	
+	int state;
+	
+	public void change() {
+		state++;
+	}
+	
+	
+
+}
