+++ TDDFeb2018/src/com/epam/tdd/lesson180216/TestFramework.java
@@ -1,0 +1,9 @@
+package com.epam.tdd.lesson180216;
+
+public class TestFramework {
+	
+	public static boolean assertEquals(int actual, int expected) {
+		return actual == expected;
+	}
+
+}
