+++ TDDFeb2018/src/com/epam/tdd/lesson180216/junit/TestModel.java
@@ -1,0 +1,14 @@
+package com.epam.tdd.lesson180216.junit;
+
+import com.epam.tdd.lesson180216.Model;
+
+public class TestModel {
+
+	/*@Test
+	public void test() {
+		Model model = new Model();
+		model.change();
+		assertEquals(model.state, 1);
+	}*/
+
+}
