+++ TDDFeb2018/src/com/epam/tdd/lesson180220/A.java
@@ -1,0 +1,18 @@
+package com.epam.tdd.lesson180220;
+
+public class A {
+	
+	int state;
+	
+//	AddTwoService add2service = AddTwoService.createAddTwoService();
+	AddTwoService add2service;
+	
+	public A(AddTwoService add2service) {
+		this.add2service = add2service;
+	}
+
+	public void doIt() {
+		state = add2service.doIt(state);
+	}
+
+}
