--- TDDFeb2018/src/lesson180216/Model.java
@@ -1,13 +1,0 @@
-package lesson180216;
-
-public class Model {
-	
-	int state;
-	
-	public void change() {
-		state++;
-	}
-	
-	
-
-}
