--- TDDFeb2018/src/lesson180216/ModelTest.java
@@ -1,13 +1,0 @@
-package lesson180216;
-
-public class ModelTest {
-
-	public void testModel() {
-		Model model = new Model();
-		
-		model.change();
-		
-		TestFramework.assertEquals(model.state, 1);
-		
-	}
-}
