--- TDDFeb2018/src/lesson180216/TestFramework.java
@@ -1,9 +1,0 @@
-package lesson180216;
-
-public class TestFramework {
-	
-	public static boolean assertEquals(int actual, int expected) {
-		return actual == expected;
-	}
-
-}
