--- TDDFeb2018/src/lesson180219/Model.java
@@ -1,20 +1,0 @@
-package lesson180219;
-
-public class Model {
-	
-	static {
-		System.err.println("Model static init");
-	}
-	
-	{
-		System.err.println("Model instance init");
-	}
-	
-	public Model() {
-		System.err.println("Model constructor");
-	}
-	
-	public void m1() {}
-	public void m2() {}
-
-}
