--- TDDFeb2018/src/lesson180220/AddTwoService.java
@@ -1,13 +1,0 @@
-package lesson180220;
-
-public interface AddTwoService {
-
-	final static AddTwoService[] service = new AddTwoService[1];
-
-	int doIt(int state);
-
-	static AddTwoService createAddTwoService() {
-		return service[0] == null ?  new SomeSlowRemoteService() : service[0];
-	}
-
-}
