--- TDDFeb2018/src/lesson180220/Fibonacci.java
@@ -1,16 +1,0 @@
-package lesson180220;
-
-public class Fibonacci {
-
-	public int fib(int n) {
-		
-		if (n < 0)
-			throw new IllegalArgumentException();
-		
-		if (n < 2)
-			return n;
-		
-		return fib(n-1) + fib(n-2);
-	}
-
-}
