+++ TDDFeb2018/test/com/epam/tdd/lesson180220/AllTests.java
@@ -1,0 +1,11 @@
+package com.epam.tdd.lesson180220;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+@RunWith(Suite.class)
+@SuiteClasses({FibonacciNegativeNumbersTest.class,
+        FibonacciParameterizedTest.class})
+public class AllTests {
+}
