--- TDDFeb2018/test/lesson180216/TestModel.java
@@ -1,19 +1,0 @@
-package lesson180216;
-
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-import lesson180216.Model;
-
-public class TestModel {
-
-	@Test
-	public void test() {
-		Model model = new Model();
-		model.change();
-		assertEquals(model.state, 1);
-	}
-
-}
