--- TDDFeb2018/test/lesson180220/AllTests.java
@@ -1,12 +1,0 @@
-package lesson180220;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ FibonacciNegativeNumbersTest.class,
-		FibonacciParameterizedTest.class })
-public class AllTests {
-
-}
